dpdk: inject local header variables into header struct instead of metadata struct
Instructions extract and lookahead in DPDK target assembler support
only operand which is a header, metadata fields are not supported
as operands.
Until now, all local variables were injected into main metadata
structure.
As a result, if there was an extract or lookahead call which
parsed the data from a packet into a header, there was generated
an instruction with metadata operand producing invalid spec file.

This commit makes a header from every local variable which is of
header type instead of adding it to main metadata structure.