ExpandLookahead: add parameter for disabling expansion of headers
Some targets (f.e. DPDK target) may support lookahead with header
type argument directly without the need to expand the header,
but they might still require expansion of structures.

ExpandLookahead pass can be utilized for such targets when a new
parameter which allows to disable expansion of headers is added.

This change is backwards compatible as a default behavior remains
the same and a new parameter for constructor is optional.